"10am. I am up, but I need to chill a bit.
Right now my despair is maximum. Really, what does it even mean to win?

Just what do I need to do to gain skill at AI? In a just world, I would have the answer, and yet in this one, a lot of my energy is focused on finding the solution.

11:05am. Let me finally start. I am going to try and finish this phase of my journey in the next few months. I need to get convex optimization and probability theory under wraps.

Right now I am filled with regret at all the mistakes and all the wasted time. And the lack of a future. I just want AI at a decent and reliable level, and yet I can't grasp it.

If I could cross this hurdle, I really could start game as it really should have been started. Even after this hurdle there will be many challenges to surmount, so why am I stuck here of all things? This place is the place I least want to be.

Give me the power to defeat the normies. And if not, the only choice is to take it myself.

11:10am. Convex optimization, probability theory - go!

On the bright side, somehow despite not properly studying set theory or the Metamath proofs, I am getting a handle on this aspect of math. Somehow or the other, rather than being a huge hurdle, set theory is turning out to be entirely trivial in practice. I could have done it without wasting nearly two weeks on Metamath if I kept a more even keel and focused on the paper more.

11:15am. Regrets, regret...

11:20am. The papers by Hart and Blackwell are ultimately just motivators for me to deal with the basics.

What must I do? What do I want?

I kind of want that Spiral with the top down type system. I really want Spiral that is easier to work in. That would really give me satisfaction. Partial evaluation might be justice, but I want more than it now.

Just think about it, me.

Deal with convex optimization. Deal with prob theory. Deal with these shitty papers.

Completely master randomized testing and get rid of the dependency on dependent type systems. Master all the aspects of mathematics that are needed with the low style. Then get back to working on the language, back to the original plan where I have the algorithm, but not the means to do proper implementation. That is where I should be.

In the future, randomized testing will be a bridge between mathematics, programming and machine learning.

I had completely underestimated it. It has the power to become this.

11:25am. Randomized testing is the alpha and the omega of programming - a lightweight formal technique when used properly.

It is it which will really allow me to close the gap between the pros and myself.

...Oh yes. There is something I wanted to say.

I've been dithering on what should and should not happen past the Singularity.

Fundamentally, the source of power should be programming itself for the Inspired.

But if you have some magic thing that allow you to do general learning, what does that leave for the programmer?

The answer is that there should be a whole family of general learning algorithms. Backpropagation + SGD is just the first batch of them.

So power and programming skill will go from manipulating discrete structures like I am doing now, to manipulating what could be described as processes. Learning processes have much in common with physical ones, and the true promise of programming lies in this direction.

I've been bashing the way deep learning research is being conducted as just throwing stuff at the wall, but there is a sliver of truth in that. There should be an essential component of randomness to not just learning, but research as well.

What matters is to find the focus. Rather than the light of inspiration being dispersed like now, if I can learn to harness and focus it, I can win.

11:35am. But what matters first is the basics.

If I can master them, then I will not be bothered by papers like Hart's ever again. Absolutely nothing will stand in my way.

...Let me just say one more thing.

Imagine a process as physical matter. So you can have a rock, or a diamond, or a lump of clay or a piece of wood.

Physical processes have very numerous attributes. They are not just described in terms of the heat coming into them. And in fact, they will have wildly different behavior to outside stimuli depending on their makeup.

In order for learning to be understood, the way it is understood needs to become more vivid and lifelike. It needs to become more physical.

I must work towards that. Right now I am in a dark bottomless pit with no way out."