"2:35pm. I am ready to resume.
Before I start let me make a remark. `{x | a^T x = b}`. I know what this set is supposed to be saying, but now that I know better, I know that set theory is a demoralization campaign made by mathematicians 70 years in advance against programmers.

I've installed Ionide and such, but there is no way I could even construct arbitrary sets that satisfy the above criterion.

Suppose I did it in Agda, how could I ever represent such a structure.

`a -> b -> \ex[ x ] (a^T x = b)`?

No way, this is just a single point. How about?

`a -> b -> x -> (a^T x = b) | not (a^T x = b)`?

Just decidability. This is such a huge issue. I can't turn arbitrary set theory terms into a program.

And something that describes a line can never be a finite set, so it must be a program.

2:45pm. Hmm, how about I represent it as a stream.

`a -> b -> stream (\ex [x] (a^T x = b))`

Maybe something like that? Some coinductive type is the only realistic way to ever tackle this in type theory. It would have to be some kind of non-repeating stream.

Set theory is always both:
- Wonderfully concise and expressive.
- Complete fiction without an ounce of rigor.

2:50pm. I just have no idea. Infinite sets are wonderful as they can be both iterated over and subsume functionality of programs in a much more expressive fashion. Yet they are fictions.

Rather than dwelling on this, let me resume the lectures.

3pm. https://youtu.be/P3W_wFZ2kUo?list=PL3940DD956CDF0622&t=850

Yeah, I am depressed right now, but I can definitely appreciate much more of the intricacy behind these definitions compared to last time.

I can do this. I'll be able to go further than before. That elipsoid definition. I've seen it plenty of times by now. It is the variance of the Gaussian distribution. I've used it plenty of times in KFAC as well.

Things often seem to rhyme in math.

3:15pm. https://youtu.be/P3W_wFZ2kUo?list=PL3940DD956CDF0622&t=1376

At around this time he says that polyhedron being convex should be proven. Ok...

I should be able to do this kind of thing now.

First of all, let me define n-polyhedron as intersection of n halfspaces.
A halfspace is a 1-polyhedron

Let me assume that a halfspace is convex.
Let me assume that the intersection of two halfspaces is convex.
Let me assume that the intersection of two halfspaces is a 2-polyhedron.
Therefore 2-polyhedron is convex.

Let me assume that intersection of a convex n-polyhedron and a halfspace is a convex n+1-polyhedron.

Therefore the intersection of a convex 2-polyhedron and a halfspace is a convex 3-polyhedron. By induction, the n-polyhedron is convex.

3:25pm. So I've covered the base case and then proved the rest by induction. I am not sure of my jargon.

3:30pm. Hmmm, this is not really formal. There is some connection missing.

But basically, this proof is intended to be a like a fold over a list.

```
let polyIntersect intersect first rest = List.fold intersect first rest
```

3:35pm. Yeah, this is the way to do this.

In Agda I'd have intersect have the type `Polyhedron n -> Polyhedron 1 -> Polyhedron (n+1)` And this would be the proof.

3:35pm. And this would be the formal proof. As can be seen, my informal proof makes too many assumptions and is less clear that the formal version. Or at least, a sketch of the formal version.

3:40pm. I am going to continue doing these sorts of exercises to show how far have I come.

In F#, I could not do the thing with dependent types, but I could set pre and post conditions for example.

```
let polyIntersect is_poly intersect first rest =
    assert (is_poly first = 1)
    assert (List.forall (fun x -> is_poly x = 1) rest)
    let r = List.fold intersect first rest
    assert (is_poly r = 1 + List.length rest)
    r
```

This could be considered a kind of proof. Now the correctness of it is bound in the implementation of List.fold and I have to assume that there aren't any side effecting computations, but assuming the computation is pure and the library functions are in order, then this is valid to an extent.

3:45pm. I need to keep doing this. I need to draw connections between proofs and programming.

4:10pm. https://youtu.be/P3W_wFZ2kUo?list=PL3940DD956CDF0622&t=2222

I hate it when they shadow like this.

4:40pm. https://youtu.be/P3W_wFZ2kUo?list=PL3940DD956CDF0622&t=3059

Ok, now I know what an orthant is.

5:20pm. https://youtu.be/kcOodzDGV4c?list=PL3940DD956CDF0622

Finally done with lecture two. This is killing me. Let me take a break.

5:45pm. Uf, let me do a little more. Though by now, my focus really is spent.

These lectures are as fun as reading a dictionary.

6:10pm. I am surfing /a/ instead of watching the lecture. This is how boring this shit is. It is amazing that there is a grand total of 37 videos of this.

...Yeah, there is no way I can deal with 3 right now. I am done for the day.

6:15pm. I am going to vow revenge for this. This idiocy needs to be rooted out from the timelines and replaced with something less retarded. Every ounce of this torturous boredom inflicted upon me will me paid back in full.

But I really am going to do it this time. 37 videos is still like 3 weeks to go through this entire monster course at most. The the prob theory book will take the same amount of time if I do not cut corners.

6:20pm. I do seem to have some motivation to go through this that I did not have before even though I am still dying from this."