"10:15am. I am up this late despite going to bed at 8:30pm. Eah. Well, let me go for Tog and then I will start the lectures.
11am. Let me start the lectures. Time for number 3.

12:30pm. https://youtu.be/lEN2xvTTr0E?list=PL3940DD956CDF0622

Let me stop here. Done with 3.

This is tedious, but by now I am decently familiar with math of this level of complexity. I can deal with it.

2:15pm. Done with chores + breakfast. Let me resume. There really is not much that is catching my interest right now so I have all the time in the world for these lectures.

2:45pm. Had to take an extra break. This is just so tedious. But I am going to do it. At least two lectures per day.

2:55pm. https://youtu.be/lEN2xvTTr0E?list=PL3940DD956CDF0622&t=1308

No seriously, enough of this. This guy just keep droning on and on about trivial bullshit.

I do not have the patience for this. It is time I start skimming. Then I'll do the prob theory book for a bit.

3pm. This course so far did not have any interesting proofs either. It literally is like reading from a dictionary. There is no way I can endure 2.5 weeks of this.

https://youtu.be/-T9cloGG_80?list=PL3940DD956CDF0622&t=1639

Even here, the same thing continues. Enough. I am going to the prob theory book.

Now that I am warming up to set theory or at least parts of it, maybe I will be able to stomach it?

3:10pm. No, actually what I need to is rethink my plan and figure out a proper way of dealing with the prerequisites to Blackwell approachability.

3:20pm. 19/548. Here it talks about infinite products of infinite spaces like its nothing.

And I put the book away never to look at it again."