"https://en.wikipedia.org/wiki/Support_function
Admittedly I did get something from the course given that I understand what hyperplanes are.

And I know what distances from the origin are.

And I know what supporting hyperplanes are.

3:55pm. I am thinking about support functions right now.

"In mathematics, the support function hA of a non-empty closed convex set A in {\displaystyle \mathbb {R} ^{n}}\mathbb {R} ^{n} describes the (signed) distances of supporting hyperplanes of A from the origin."

This is what Wikipedia says the support function is.

The thing is, this does not at all match what the definition is saying: `h_A(x) = sup{x*a : a in A}`

How is this a signed distance from the origin? A signed distance would be something like a norm...

4:10pm. https://math.stackexchange.com/questions/235926/what-is-a-support-function-sup-z-in-k-langle-z-x-rangle

Agh, I need to think about this.

https://www.wolframalpha.com/input/?i=support+function

Nope...WA does not recognize set theory.

4:45pm. Right now I've moved to the paper. It is not necessarily true that I am not learning anything, but what I am learning is not necessarily the theorem itself.

Let me take a look at the proof of the main theorem more carefully and I will gather my thoughts.

5:05pm. 19/32. The steps here are so ridiculous.

5:15pm. Let me go for the paper by Blackwell again. I am starting to come to a resolution.

I want to do all these steps. I mean of course...if I continue learning for the rest of my life, I still won't be able to do them. Because they are all fluff. That papers like this smell, I did not really need over 6 months of theorem proving experience to conclude.

But even if I can't win. Even if I am destined to fail...there is a point in going through the definitions in putting them in terms of functional programming. I can never do more than create a simulacrum of the real thing, but the fact that I can do at least much right now should be a point of pride.

I should be able to tackle papers like these with courage. If nothing else, I should be able to separate the computable from the non-computable.

I will never get the ability to create these vacuous proofs, but I sure should be able to give them a good shake.

If this kind of math is a middle finger to computability, then my job should be to reach out and snap it.

5:20pm. I should not avert my eyes from this wretched mess that is modern mathematics. Instead I need to aim to scavenge it.

5:50pm. Had to take a break. Let me put in one last spurt of effort into the paper by Blackwell and then I will call it a day.

```
Prob {d n >= e for some n >= N} < e
```

This completely baffled me last time, but here is what I think it means.

```
{d n >= e | n >= N} | / | { n | n >= N} | < e
```

6:15pm. https://www.youtube.com/watch?v=iQe_x_dhSE8
Command & Conquer Red Alert OST - Face The Enemy 1 (High Quality)

https://www.youtube.com/watch?v=0j30rWyvglo
Command and Conquer- Red Alert 1 - Face The Enemy 2

These songs are amazing. Back when I played Red Alert I wasn't at an age where I'd pay attention to the OST. I considered it the part of the background. But now that I am more mature, I see just how good they really were.

6:40pm. I finished lunch 5m ago, but let me hero these papers a little more.

6:45pm. I will call it a day here. I am tired at this point.

These two papers are essentially bereft of computational content, but that is my challenge in this era. How close is the closest I can get to turning them into something sensible?

It is not like the papers are nothing - there should be enough info for at least the definitions.

And that is what I must aim for starting from here.

6:50pm. There is no doubt, that my behavior here in the face of these difficulties is nothing short of heroic, but I won't be able to break through in the direction that I want.

So I'll develop that essential element of style - the skeleton coding.

I have to be more realistic no matter what.

Sure the most realistic thing would have been to go with Python back in 2015, but training is important too. I'll make my cut on these papers and then go back to programming. By that I mean implementing the agents in their full glory, and using the reversibility idea for stability in hierarchies.

Since the GPU is not that useful for RL, I'll use F# for everything. Eventually I will get back to Spiral, but getting the implementation at full power is not what really matters now.

I've been too much in a hurry to create Spiral.

The cleaning up that I am doing now is the right thing to be doing."