"9:55am. https://www.youtube.com/watch?v=YZ-_E7A3V2w David Duvenaud | Reflecting on Neural ODEs | NeurIPS 2019
I'll leave this for later.

https://openreview.net/forum?id=SklKcJ3EFH
Newtonian Monte Carlo: a second-order gradient method for speeding up MCMC

10am. Right now I am groggy. Let me chill a bit before I start. What I want today is make a dent into those two papers.

What I will be doing now will be the only time in my life I will do this kind of thing so I might as well do it.

I've come to terms with it - my talent definitely does not extend to the non-computable. Math that talks about infinities and juggles it is not something I can deal with it. Math that belongs in the fiction section is not something I will ever become good at. Math that pretends at rigor is not my cup of tea.

But while I say that, actually proofs are hellishly difficult even with my knack for programming, and I'd need to be a superhuman intelligence to really correct the injustices of the 20th century in this field.

10:05am. It was not always like this. Math went wrong only in the last 100 years, and the kind of math that is computable was the norm before that.

So probably what used to be mathematicians then would have been programmers today.

Today, the dividing line between mathematics and programming seems to be the attitude towards computability.

10:10am. I'll do this for a week longer and then get back to my original plan which is to go through Lanctot's thesis.

10:25am. Auh, I really do envy the Inspired. Imagine doing everything on your own, from theory to the libraries, even that hardware and the OS the processes will be running. With a brain cores, what would take me an entire lifetime, they could do in a few days or less. Nobody could possibly compete with that.

For humans talent is a constant, but for them talent is a variable that can be optimized. Talent is like a muscle - it is not something that can be trained if one does not have it. They will have it in endless abundance.

I feel that math being computable would be justice, but that is cold comfort in this crazy era, this clown world.

Let me take a look at the Newtonian MCMC paper and then I will start work on the paper by Hart.

Damn it, even if I can't figure it out, I sure want to give it a good punch."