"10:45am. Let me start this thing. Let me load up the paper by Hart and then I will go through the definitions mechanically.
10:55am.

```
// 3.1
let qR_eq g q t i j =
    sum' i g (fun k ->  q t i k * R g i t k j) = q t i j * sum' i g (fun k -> R g t i j k)

let Blackwell_approachability dot i l v q s w C =
    dot l (v (fun i' -> if i = i' then q l else s i')) = w C l
```

This is really hard to bear.

But let me see how far I can go with it.

11:10am. Not very far apparently. This is hopeless.

I can't gather any motivation when I know that this will never be live. It is just demoralizing to try to write these things down when I know they can't have any content like this.

And I can't really transform this into something that makes sense at my level.

11:35am. I am going to the next phase of my plan.

It is just too difficult emotionally for me to do this.

I am not really excelling here, I can't deny that.

I might not be able to reach my goals here, but I have received a great deal of experience in how set theory is really used in the real world. And a good lesson in how important computability really is.

To close the story on the papers by Hart and Blackwell, let me just say that despite the stuff in them being fiction, having either Blackwell or Hart as mentors and having them explain to me their thinking step by step would be an immense aid in understanding this.

Since the theorems are uncomputable, the computers are not helpful here as they otherwise would be.

11:40am. I cannot get to the truth of them by implementing and playing with them.

The algorithms themselves will have to be their own proofs for the time being.

11:45am. That having said, I am not completely done yet.

There are references by Lanctot I've yet to look at.

And it is true that my ability to understand these papers has improved, though not to the point where I can understand this properly.

The plan here will be to go from the top through the references until I arrive back at his thesis.

Let me start that now. I'll do it for 30-60m and then have breakfast."