"10:05am. Programs (that terminate) are proofs and computation is the one true language. The mathematicians will pay for turning their back on it.
For the sake of the argument suppose this 100 years into the future and some version of Agda is the most popular language. Suppose you look at some library or big piece of code.

The first thing you'd do is look at the proofs.

You certainly wouldn't start with the algorithm implementations themselves. Because those are hard.

10:10am. Today if you go on the ML sub, you never see discussion of proofs. Always the proof of validity is the code itself. Not the proofs in the paper. Plenty of complaints if the paper does not come with code included, but nobody except the editorial boards care about proofs.

It is finally time to accept this - there is something wrong with mathematical practice. Fundamentally, proofs of programs should be simpler to understand than their core parts. Fundamentally, proofs should be about understanding.

This stuff in these books and papers that I am trying to digest - just what the fuck is this? I am really, really going out of my way to get it and I am not at all successful.

All these issues with non-computability, all these steps that seemingly come out of nowhere that I could only figure out by picking the author's brain. Just what is the point of even learning math if I have to be a literal mind reader to get it?

10:15am. Yes, this is hard.

But what I am trying to say is, it definitely should not be.

A proof should be an elegant explanation of some property - its purpose for existence should not be to show that the author has a 200 IQ. It should not be encrypted code for 5 other people in the world who understand it.

10:20am. Mathematics as it is done today can just die. There is no value to it. It went wrong in the 20th century and it is up to computer science to fix it.

I admit I had been looking down on webdev coders and Pythonistas as brainlets. But compared to mathematicians any programmer is a bastion of rigor. Mathematics should be eaten by computer science.

Eventually crappy languages will die out and what will be left is the real deal.

10:25am. But forget looking 100 years in the future and the hypotheticals. The Singularity needs to happen before that. It is not enough that math should die, I need to kill it myself. Because I do not have the time to wait 100 years.

I need to act.

Rather than changing plans, this is more like me going back to the original plan. I've been bashing the field of ML's ad-hoc way of doing things in the recent reviews, but maybe that is the only choice left at this point.

10:30am. I have two goals as usual.

1) Level up my programming skill.
2) Make money along the way.

I've been pretty successful at the first, at the expense of the second. I really must get on with the plan.

10:30am. For poker I just need to...

1) Make the agent.
2) Use the agent.

Forget implementing Lanctot's thesis, and the stuff in the newer papers for the time being. Just as a mental exercise, suppose I get Deepstack made in Python from somewhere. What is stopping me from just using it? It is not like I need to spend a few months implementing everything on my own.

10:35am. This is the starting point I need to take in my thinking. Rather then trying to solve the ML library implementation problem with my own special language, it makes more sense to start from a position of abundance.

I am not one of the Inspired. I should not pretend that I can build up all the infrastructure that is needed on my own. The infrastructure is a constant amount of effort, and it is doable, but as a human I really do not have the time for it.

10:40am. I am going to start taking ML realistically. That means in this era starting with Python and getting pre-canned players.

This way of doing things and those players will no doubt have problems, but it is those problems that should be my problems. I should not be doing things from the ground up. Otherwise what will happen is just what happened - I'll have spent 5 years with nothing to show for in this area.

Well, that is what I wanted to say. Maybe I will need to make my own players in F# or whatever. It depends on what is out there. I might need to implement the papers on my own.

Of course, for the sake of leveling up my programming skill, I am going to have to make Spiral v0.2 at some point. I am going to have to do more theorem proving in Agda, and go through Wildberger's vids in order to brush up my math fundamentals.

But for the time being, let me aim to lessen the workload on me as much as possible. Since the situation with math is preventing me from standing out, I might as well blend in for just a bit.

10:45am. So what is my priority from here?

Let me just take a few peeks at the links I posted yesterday.

Then I will forget about strategy averaging and start with Deepstack. I'll see whether anything new came from that line of research and try hunting down the code and trying it out. Otherwise, I will look towards implementing something myself.

That is the way things should go. Assuming I find something good in a can, I should just move to part two which is making the interfaces for those online poker rooms.

After that...well, I can always learn math or start work on Spiral v0.2 while the agents are actually active in the background.

10:50am. I want to start this right away.

11am. Let me do it. I'll take a look at the code for 30-60m and then have breakfast."