"11:05am. https://github.com/deepmind/open_spiel/blob/e6eb15708c722b3e61ef7bed3a93a2bbbfd0900c/open_spiel/algorithms/matrix_game_utils.cc#L92
This code is really quite difficult. As expected C++ is not F# or Haskell or Agda. You can't really study this.

Nevermind this. As expected unless I want to compile this all on my own and trace through it, I am wasting my time.

What is the next thing on the list?

Let me look up citations for Deepstack.

https://arxiv.org/abs/1908.09453
OpenSpiel - A Framework for Reinforcement Learning in Games

Lanctot is the first author on this out of many.

Yeah, I think OpenSpiel is definitely something I will be studying.

"Running any of the DeepStack code requires Lua and torch. Torch is only officially supported for *NIX based systems (i.e. Linux and Mac OS X), and as such, we don't officially support installation of DeepStack on Windows systems. This page contains suggestions for using torch on a Windows machine; if you decide to try this, we cannot offer any help."

Lua and torch. So it is not even Python and PyTorch?

And the implementation out there is just for Leduc poker.

11:15am. Right now I am going through the citations for Deepstack from 2019

http://proceedings.mlr.press/v97/farina19b.html
Regret Circuits - Composability of Regret Minimizers

I'll save this for later.

I am trying to come up with a plan. I am not sure I really want to program in Lua. And I really do not want to switch OSes.

So I will be doing it on my own.

Ah yes, ah yes. I completely forgot about it by this point, but there was a Deep CFR paper wasn't it?

That is what I should be looking into.

It is no problem at all. I'll bring in the .NET lin alg library and implement a small NN on my own. That is really all I need here. 3 layers feedforward should be more than enough.

12:20pm. https://mangadex.org/title/37415/samayoeru-tensei-sha-tachi-no-relive-game

Since I am having breakfast, let me commit here. This manga is pretty good. I've seen it in threads before, but it is my first time reading it.

In the past few days I did not have much time to have fun due to my distress, so now that the pressure has let off on me, I can cool down a little.

12:25pm. Right now my current plan is Deep CFR, DeepStack and OpenSpiel papers. Then I can start making plans for actually implementing the things in it."