"2:55pm. Ah, let me take a break. I really will need some time to really get into this. I am shifting my whole mindset and that is never easy.
3:30pm. https://github.com/deepmind/open_spiel/blob/master/docs/windows.md

No GPU support, can only be run using a virtual machine.

3:40pm. https://github.com/deepmind/open_spiel/blob/master/open_spiel/python/algorithms/cfr.py

In order to make progress all I need to do is to simply decide to start.

I do not need to look for the best algorithm.

https://github.com/deepmind/open_spiel/blob/master/open_spiel/python/algorithms/deep_cfr.py

This should suffice. Now, how can they make it work for poker without recurrence?

3:55pm. Had to take a short break. Let me resume.

The most important thing in getting things done is not so much the language or the libraries. It is having a clear determination and a goal.

The reason why the things I done up to now have taken years is because all the wandering.

If I make it my goal to just implement Deep CFR + HU NL Holden it won't take that long. I certainly do not feel like switching to Linux + Python + C++. It might be a tad suboptimal, but let me just do what needs to be done in F#. F# is a the ideal compromise for me between emigrating completely and starting work on Spiral v0.2.

Very well.

Right now my goal with be to implement Deep CFR in F#. Work on this will start immediately. I will ignore reading the papers about the alternatives and focus only on getting it done.

Let me study the Python file. Obviously I won't be starting work on it today as I still need to get over my depression, but I am going to study the paper and the Python file I just linked to.

What I need to do right now if practice swinging the sword in my head. Then I'll do it in real life.

That is how all this goes.

5:50pm. I sure spent a lot of while thinking this through.

I've come to a conclusion. This is the ideal time to do Spiral v0.2.

5:55pm. I mean yes, sure, I can do Deep CFR in Python, F# or whatever.

But you know what I could do that would make me a lot more money with a lot less variance? Get a job.

I just can't see anything interesting at all in doing Deep CFR. I can do it. It is no problem.

But I am having all these thoughts that it is time to make money, but at the other end I have all these ideas for Spiral v0.2 that I want to try out and yet I am reluctant because it will take forever just like v0.09 did.

But I am sure. If I get a job here or if I start messing with RL in Python, I will never get the chance to make v0.2 of Spiral. If I want to get it done, I should strike while the iron is hot.

6pm. If it could not be such a dead end financially, I'd very much like to try implementing the CFR algorithms and maybe a new ML library while working on a new version of Spiral.

https://www.patreon.com/

So if I am going to gamble, why not test out whether there was any point to all that Reddit posting that I did on the PL sub. Spiral has 700 stars at the moment. I have no idea whether that is a lot or a little, but a few people have heard of me.

Honestly, if I can get enough people to pitch in for the cost of buying a new computer (2k), I would be quite happy. Getting something like a yearly salary where I live (15k) would make me ecstatic.

6:05pm. It is not right to drop Spiral right here in the middle. It is my darling.

I have to crazy. I have to be insane. I know it makes zero sense to approach making money through RL the way I am doing it.

6:10pm. Yet I do not want to be forced into Python. I do not want to program the way circumstances impose upon me.

I want to develop my own style and my own language. I want to continue my exploration for just a while longer.

Just like I did the bottom-up in Spiral, I want to try the top-down too and see where that gets me.

6:45pm. Done with lunch. I'll call it a day here.

Even I started now I do not think I could set anything up on Patreon in just two weeks. I want some time to prepare and think about it first.

6:55pm. I really like the style F# propagates, but at as a language it just lacks all the necessary prerequisites. It is not sufficiently universal. For a language to be universal, it should be C tier efficient while at the same time having great capability for abstraction. Without the former you cannot have things like GPU backends, and without the later it is not worth using.

7pm. Ultimately, all I want is higher order logic together with partial evaluation. The language Spiral could become would be enough to satisfy me.

Though I have regrets, I really was great during 2017-2018. Working on languages like this for some far off goal is the truest to my character.

7:10pm. I should be doing what I was doing. I should not give up on my plan just because things are not going my way.

I'll implement Deep CFR, but it will be on my own terms. It is too bad about Simulacrum, but programming is what I should be doing. The level of absolute control that I seek is still something that I hadn't achieved.

If you are going to pursue something, you should not make compromises.

If I were Inspired, I would do it like this. No way would I be satisfied using somebody else's language.

The field of ML will take care of itself eventually.

I have only one life and what I should be doing is working on a language."