"4:10pm. As an exercise, it might be worth implementing Deep CFR in Haskell.
4:30pm. Ah, forget it. How about I call it a day here? Right now I am just surfing 4chan randomly. It is a great waste of time. There haven't been any good threads for a while now. I really liked that comp sci post on /sci/, but I've lurking it for days now and literally everything thread was shit.

There is nothing to it, but to go for it.

I'll try to see whether I can do some programming tomorrow.

4:35pm. To think I would be coming back to Haskell willingly. I am amazed.

Let me just say, that if I in any way believed that the algorithms could really deliver, I would drop everything and do it all in Spiral or whatever.

Maybe the fact that I am taking these choices is reflective of how I truly feel about the current state of ML.

Well then...if I really believe in the future, then I should just work on my programming skills.

None of these algorithms are really that complicated. This very much goes for DL.

But the languages not being developed enough means that what would otherwise be easy is now hard.

4:40pm. I won't blame myself for this state of affairs.

I've chosen this path. If I had powers beyond that of humanity, I could have competed with others for real rather than just fight against myself, but the proof of my existence is this desire.

Let's master it all.

Unlike Spiral, I do not strictly speaking need to master standard FP, but it won't hurt."