"1:50pm. I actually have neck pain from all the bad sleep posture due to tension.
I really need a break, but I doubt this obsession will leave me just like that.

2:15pm. I think today I will aim to rest for a while.

I want to think about these ideas that I have some more, but I really should get some Haskell experience after that.

4:25pm. Right now I am busting my head trying to figure out how to deal with return type polymorphism.

I am thinking things through and have realized that the partial evaluator does not have enough information to make the decision at every point regarding which typeclass instance to select.

```
let f (x : 'a) : 'c =
    let q : 'b = g x
    let _ = 1 + q
    h q
```

I am supposing that `g` is some typeclass here.

The issue is that 1.

I do not have a mechanism to tell it that it is the same type as `q` here.

In fact, had I written it like this `1 + g x`, the actual type does not even exist till I extract it by running `g x`.

4:30pm. Tricky. Tricky!

https://en.wikipedia.org/wiki/PCP_theorem
"In computational complexity theory, the PCP theorem (also known as the PCP characterization theorem) states that every decision problem in the NP complexity class has probabilistically checkable proofs (proofs that can be checked by a randomized algorithm) of constant query complexity and logarithmic randomness complexity (uses a logarithmic number of random bits).

The PCP theorem says that for some universal constant K, for every n, any mathematical proof of length n can be rewritten as a different proof of length poly(n) that is formally verifiable with 99% accuracy by a randomized algorithm that inspects only K letters of that proof.

The PCP theorem is the cornerstone of the theory of computational hardness of approximation, which investigates the inherent difficulty in designing efficient approximation algorithms for various optimization problems. It has been described by Ingo Wegener as "the most important result in complexity theory since Cook's theorem"[1] and by Oded Goldreich as "a culmination of a sequence of impressive works […] rich in innovative ideas"."

Oh this is interesting.

Somebody mentioned something good on /sci/.

Never heard about this before

8:15pm. I know I said I would be resting, but those brief moments of respite did not last long. Right now I am so focused on the issues at hand that I might bore a hole through reality. My mind feels feverish.

As it turns out there are all sorts of issues that you only realize once you really focus on a subject.

Actually, up to now I had not realized at all why a single top down pass is not enough for typeclasses or why dictionary passing would be necessary at runtime.

8:20pm. Return type polymorphism is just the start of the problems. As soon as I realized that, pretty soon I could see that I would need to do unification during partial evaluation.

Then I realized I would have all sort of scoping issues with metavariables and such. Not to mention having to deal with them in the partial evaluator itself.

8:25pm. Right now I am thinking that just collecting all the constraints and resolving them at the function call site would work the best. It would be the simplest way of doing things. And as a bonus since I would just be solving constraints, there would not be a noticeable difference between return type polymorphism and the regular kind.

8:30pm. Let me go to the previous example.

```
let f (x : 'a) : 'b = h (g x)
```

Here `f` should not be necessarily seen as `f : 'a -> 'c`. If I was passing `g` and `h` by hand it would be `f : 'a -> ('a -> 'b) -> ('b -> 'c) -> 'c`.

I've been thinking this function as having hidden extra methods passed via the typeclass dictionary, but it is not just that. It has hidden foralls as well.

8:35pm. That is a key thing I need to focus on.

Rather than me somehow doing unification during the partial evaluation pass, it would be by far the easiest for me to do as much analysis as I can during typechecking and then resolve all the difficulties with ad-hoc polymorphism with some constraint solving that I can put into a single function.

8:50pm. Now I am thinking how to order the unification search. It will necessarily have to be unification with backtracking like in Prolog. But it won't be too hard.

I am going to break from tradition.

```
f : forall a b c. Constraint1 a, Constraint2 a b, Constraint3 a b c => a -> b -> c
```

Will be...

```
f : forall a. Constraint1 a => a -> forall b. Constraint2 a b => b -> forall c. Constraint3 a b c => c
```

I am going to infer the types to be as general as possible.

Also inlineables will really come in handy when it comes to passing higher ranked types to and from functions.

They will make things more efficient and enable symmetry.

9:05pm. For large nested functions like those I like to use in Spiral, this will lead to a lot of free variables depending on the number of typeclass instances being applied. Also if I do call site solving how am I going to make autogenerated instances for things like records? How will I use typeclasses from inside raw Spiral?

9:10pm. Ah, enough. Let me shut down here. Rather than just sitting in this seat while I think, I might as well do that in bed after I take a bath.

This really is making me exercise my brain like nothing else.

I am back in the good old days of PL design."